fix(device_handle): Allow SET_IDLE requests with direction out
The SET_IDLE request (bRequest 0x0a) have a bmRequestType of 0x21. Using
this request_type in libusb-rs DeviceHandle::read_control gets rejected
with an InvalidParam, because it's direction is not Direction::In, but
rather Direction::Out (when parsed as direction).

This Pull Request adds another condition to the request_type Mask-Check,
allowing SET_IDLE requests of `request == 0x0a` to pass through the check
with Direction::Out set.